Move log types
