+++ src/types/log.types.ts
@@ -1,3 +1,5 @@
+import { FileType } from './associatedArtifact.types';
+import { LogEntryObject } from './log.types';
 import { Message, TicketType } from './publication_types/ticket.types';
 
 export interface Log {
@@ -27,4 +29,57 @@ export interface LogActionItem {
   description: string;
   date?: string;
   fileIcon?: 'file' | 'deletedFile' | 'archivedFile' | 'rejectedFile' | 'hiddenFile';
+}
+
+interface LogEntryOnBehalfOf {
+  id: string;
+  topLevelOrgCristinId: string;
+  shortName: string;
+  displayName: string;
+}
+
+interface LogEntryPerformedBy {
+  givenName: string;
+  familyName: string;
+  cristinId: string;
+  organization: string;
+  onBehalfOf: LogEntryOnBehalfOf;
+}
+
+interface BaseLogEntry {
+  timestamp: string;
+  performedBy?: LogEntryPerformedBy;
+}
+
+interface PublicationLogEntry extends BaseLogEntry {
+  type: 'PublicationLogEntry';
+  topic:
+    | 'PublicationCreated'
+    | 'PublicationPublished'
+    | 'PublicationUnpublished'
+    | 'PublicationDeleted'
+    | 'PublicationRepublished';
+}
+
+interface PublicationImportedLogEntry extends Omit<PublicationLogEntry, 'topic'> {
+  topic: 'PublicationImported';
+  source: {
+    importSource: {
+      archive?: string;
+      source: string;
+    };
+  };
+}
+
+interface FileLogEntry extends BaseLogEntry {
+  type: 'FileLogEntry';
+  topic: 'FileUploaded' | 'FileApproved' | 'FileRejected' | 'FileDeleted';
+  filename: string;
+  fileType: FileType;
+}
+
+export type LogEntryObject = PublicationLogEntry | FileLogEntry | PublicationImportedLogEntry;
+
+export interface RegistrationLogResponse {
+  logEntries: LogEntryObject[];
 }
